test: restructure suite into unit/integration + extract mocks
Structural half of the test-suite cleanup:

tests/mocks/ package grouped by domain (flow_data, training, caching),
registered via pytest_plugins. conftest.py keeps only the collection
hooks and the cross-cutting clean_tmp_path fixture.

src/flowgym: common, flow, density, nn, training) and tests/integration/
(cross-module + end-to-end: training pipelines, eval/estimator caching).
Files are moved, not renamed, so the basename-based other_methods gating
in conftest still applies. Switches pytest to importlib import mode and
adds 'tests' to pythonpath. Adds tests/README.md documenting the layout.

Per-test docstring improvements follow in a separate commit.

Relates to #21, #22
