+++ tests/integration/training/test_full_integration.py
