+++ tests/integration/training/test_integrated_checkpointing.py
