+++ tests/unit/common/base/test_base_cache.py
