+++ tests/unit/common/test_data_smoothing.py
