+++ tests/unit/density/test_density_estimate_simple.py
