+++ tests/unit/nn/test_registry.py
