+++ tests/unit/training/caching/test_cache_key_generation.py
