+++ tests/unit/training/caching/test_caching_rules.py
