+++ tests/unit/training/caching/test_generic_caching.py
